Move looks_like_adduct to derive_adduct_from_name.py
