+++ matchms/filtering/repair_adduct/clean_adduct.py
@@ -5,22 +5,6 @@ from functools import lru_cache
 from typing import Dict
 
 
-def looks_like_adduct(adduct):
-    """Return True if input string has expected format of an adduct."""
-    if not isinstance(adduct, str):
-        return False
-    # Clean adduct
-    adduct = clean_adduct(adduct)
-    # Load lists of default known adducts
-    known_adducts = load_adducts_dict()
-    if adduct in known_adducts:
-        return True
-
-    # Expect format like: "[2M-H]" or "[2M+Na]+"
-    regexp1 = r"^\[(([0-4]M)|(M[0-9])|(M))((Br)|(Br81)|(Cl)|(Cl37)|(S)){0,}[+-][A-Z0-9\+\-\(\)aglire]{1,}[\]0-4+-]{1,4}"
-    return re.search(regexp1, adduct) is not None
-
-
 def clean_adduct(adduct: str) -> str:
     """Clean adduct and make it consistent in style.
     Will transform adduct strings of type 'M+H+' to '[M+H]+'.
@@ -31,7 +15,7 @@ def clean_adduct(adduct: str) -> str:
         Input adduct string to be cleaned/edited.
     """
     def get_adduct_charge(adduct):
-        regex_charges = r"[1-3]{0,1}[+,-]{1,2}$"
+        regex_charges = r"[1-3]{0,1}[+-]{1,2}$"
         match = re.search(regex_charges, adduct)
         if match:
             return match.group(0)
